Conflicts:
	app/config/security.yml
	src/Pelagos/Bundle/AppBundle/Security/ResearchGroupVoter.php

+++ src/Pelagos/Bundle/AppBundle/Handler/EntityHandler.php
@@ -138,10 +138,17 @@ class EntityHandler
      *
      * @param Entity $entity The entity object to delete.
      *
+     * @throws AccessDeniedException When the user does not have sufficient privileges to delete the entity.
+     *
      * @return Entity The entity object that was deleted.
      */
     public function delete(Entity $entity)
     {
+        if (!$this->authorizationChecker->isGranted(PelagosEntityVoter::CAN_DELETE, $entity)) {
+            throw new AccessDeniedException(
+                'You do not have sufficient privileges to delete this ' . $entity::FRIENDLY_NAME . '.'
+            );
+        }
         $this->entityManager->remove($entity);
         $this->entityManager->flush();
         return $entity;
